[python] Improving code qualityi for fast Reindexer
