--- apis/python/src/tiledbsoma/utils/__init__.py
@@ -1,5 +1,0 @@
-from . import IntIndexer
-
-__all__ = [
-    "IntIndexer",
-]
